Fixed a bug in single line tabbing that caused the tab length not being taken into account.
Added DOM emulation to the unit tests so that they
can be run from node.

Extended the Mockrenderer for the unit tests.

Made the majority of the unit tests work again
(barring those for moveLinesUp, see below).

Fixed a bug in moveLinesUp that caused an empty
line to be inserted. However, this fix causes the
editor to break (need to look at this).