Fixed gotoLine to consider wrapped lines when calculating where to scroll to
I have created a new method VirtualRenderer#scrollToLine to do for lines what
scrollToRow does for absolute rows in the editor. Previously gotoLine called
scrollToRow which would jump to the nth row, but with line wraps this may not
be the nth line. scrollToLine also takes an additional parameter to tell it
whether to scroll the line to the top or center of the editor.