removed left property assignment
When the editor takes up the entire window and the window is only half
the size of my screen, the menus get crunched up horribly.  I'll provide
for closing the overlay by pressing esc in case the overlay's contents
take up the entire screen.

If ace is to be used in very small frames then implementors will need to
provide styling information in css to control how the settings menu
looks.