adds drop_load extension
I pulled the drag and drop code out of the kitchen sink demo and
augmented it a bit. It accepts a callback so that the editor should load
the contents of the dropped file and then call the callback with the
`file` and `FileReader` objects for further processing. If no callback
is specified it defaults to a no op function.