Proposal for html completion
I saw there was an issue about html tag completion, suggesting a
behavior like zencoding, and as I've been working with ace for a project
with an unusual xml schema, I thought I could collaborate integrating
some html completion (and the backbone for future xml completions) with
the current code structure.

Basically, I'm adding an insertion behavior for XML that wraps the
current completion with the characters and attributes specified in an
object literal very like the keywords variable in the modes' highlights.

When you start typing a tag, like `<link`, you can press `control+space`
to open up the completion list, it will appear with normal words and
words that start with `<` (pointing that they're elements), once you
pick a word (for this example: `<link`) it will autocomplete the next
code:

    <link href="" rel="stylesheet" type="text/css" title="">

It also works if you type `<link` and then close the tab without
oppening the autocomplete dialog, so, for example, when you type `<a>`
it will automatically generate the full _a_ line, as follows:

    <a href="" rel="" target="_blank"></a>

It will always put the cursor within the tags (after `>` and before
`</`).

It can be improved, this is just a first sketch, what do you think?