Document : small correctness and performance improvement to `getAllLines` function
Previously, this function was wrapping a call to `getLines(0,
this.getLength())`, but the `getLines` uses an inclusive `lastRow` and
adds 1 before calling `slice` on the array.  So the `end` argument was
actually 1 greater than the length of the array.  In practice, this is
clipped? and does not cause any damage.  But this change is more
correct, and should perform better as it eliminates 2 calls (getLength()
and getLines()) and 1 increment (lastRow + 1).