Add missing `$id` to those Modes that are missing it.
I've found that the set of modes in this PR did not set their `$id`.
The side-effect of that is that when one of these modes is the current
mode of an EditSession, and one calls `session.getOption('mode')` the
returned value was incorrect.  It goes up the prototype chain so you'd
get something like 'ace/mode/text' or 'ace/mode/html' in most cases.

This fixes that so that if, for example, the mode is set to
'ace/mode/django', the value returned by the option will be
'ace/mode/django'