Change Vim emulation to use long-word selection.
Key bindings like i_<C-W> should delete through whitespace. While this change does not even begin to emulate Vim's selection behaviour, it should at least ameliorate it for now. 

This patch is unnecessary as it doesn't quite resolve this misbehaviour. This change will likely have to be removed in favor of a Vim-specific selection strategy.