Renamed the theme kr_theme to kr
This is in line with the other templates and creates a no conflict
version of the theme which is under the namespace ace/theme/kr rather
than ace/theme/kr_theme. This is important because during the build process
theme is stripped from the file name. So the kr.js file has kr_theme in it.