mode/rst: fix some regexps to comply with standard rst
The official reStructuredText parser from the docutils library has some
requirements to determine if a "**" sequence is the start of inline bold
text.

- The character before "**" must be either:
  + The begining of line
  + A space
  + One of " ' ( < [ { - / :
- The next character after "**" must not be a space.

Then to determine if a "**" is the end of inline bold text.

- The character before "**" must not be a space.
- The character after "**" must be either:
  + The end of line
  + A space
  + One of \ . , ; ! ? - / : " ' ) > ] }

The same requirements apply for inline italics "*", literals "``" and
interpreted role text "`".

See rst parser source code for more details:
http://repo.or.cz/docutils.git/blob/HEAD:/docutils/docutils/parsers/rst/states.py#l527

Signed-off-by: Robin Jarry <robin.jarry@6wind.com>