mode/rst: fix rst demo
The example taken from
http://docutils.sourceforge.net/docs/user/rst/quickstart.html contains
code blocks nested into definition lists.

This is not supported by the rst mode tokenizer as it would require to
memorize the initial indentation level when the code block starts to be
able to determine whether it ends or not. For now, the tokenizer ends a
code block (or directive body) when it encounters the first line without
indentation.

Fix the example by unindenting the definition lists containing code
blocks.

Also, fix the top heading by adding an overline.

Signed-off-by: Robin Jarry <robin.jarry@6wind.com>