ext-emmet: Adopt newer emmet's API
emmet had dropped their own `define` and `require` at v1.1.
`emmet.require` is no longer available and they changed the way to export API.

The following shows the changes over API exports:
* emmet.require("resources"): emmet.resources
* emmet.require("tabStops"): emmet.tabStops
* emmet.require("utils").replaceSubstring:  emmet.utils.common.replaceSubstring
* emmet.require("actions"): emmet.actions

The usage of emmet.require("range") is dropped since this is not public and
replaceSubstring works fine without range.

Note that the backward compatibility is strictly maintained.
This change can be used with older releases of emmet.