save + restore state when using heredocs
This should fix an issue where modes that embed `sh` highlight rules can get into an invalid state after entering and closing a heredoc.

The existing code seems to already attempt to restore an old saved state; however, as far as I can tell the state is never actually 'saved' when the heredoc is entered. That seems unintentional to me but it's possible I'm missing something?