Java: Add Java 9 module-related keywords
Registered the ten new Java 9 restricted keywords as keywords.
From Section 3.9 of http://cr.openjdk.java.net/~mr/jigsaw/spec/java-se-9-jls-diffs.pdf:
> A further ten character sequences are restricted keywords: open, module,
requires, transitive, exports, opens, to, uses, provides, and with.

Note that these are *restricted* keywords, it means that ideally they should be highlighted as keywords only when in `module-info.java`. From the Section 3.9 referenced above:
> These
character sequences are tokenized as keywords solely where they appear as
terminals in the ModuleDeclaration and ModuleDirective productions (§7.7). They
are tokenized as identifiers everywhere else, for compatibility with programs
written prior to Java SE 9. There is one exception: immediately to the right of
the character sequence requires in the ModuleDirective production, the character
sequence transitive is tokenized as a keyword unless it is followed by a separator,
in which case it is tokenized as an identifier.

It makes this change similar to https://github.com/ajaxorg/ace/pull/3467 where `var` is not a keyword, but a `reserved type name`