Allow completions with empty values
Hi, Ace team! My team has been using a fork of Ace, and I'm hoping to
merge some of our changes upstream so that we can return to using your
official releases.

The engineer who originally accomplished this work has long since
departed my team, so please forgive this somewhat rote transcription of
his working notes.

```
Previously, completions with an empty string for `value` would display
correctly as a suggestion in the popup, but would error trying to
`Document.$detectNewLine` on the completion object if the user tried to
insert that suggestion.

Autocomplete now respects empty string as a value, doing an insert of
nothing on select that suggestion (rather than erroring).  This is handy
to properly support since it enables use cases like popping up a hint
in place of completions in cases where no suggestions are possible.
```