Allow user feedback for empty autocompletes
Hi, Ace team! My team has been using a fork of Ace, and I'm hoping to
merge some of our changes upstream so that we can return to using your
official releases.

The engineer who originally accomplished this work has long since
departed my team, so please forgive this somewhat rote transcription of
his working notes.

```
Add support for setting a completer.emptyMessage function which
generates a message to display in the completions popup when there
are no completions.

This enables configuring the editor to always give the user feedback
when they ask for suggestions, even if it is just to note that there
are none.
```