Add snippet support for LaTeX and custom macro completer.
The snippets are copied from the snippets for TeX.
The current keywordCompleter for LaTeX is unused because the highlight rules
do not give a list of keywords, but color every LaTeX macro, so the completer
cannot use the rules to provide completion. This macro completer is inspired
by the text completer, but autocomplete only macros (the current text completer
do not consider \ to be a part of a word).

The interface of Completer is changed so that the keywordCompleter can forward
the triggerCharacters and identifiersRegexps of `session.$mode.completer`.